requirements.txt
openenv-core==0.2.1
openai>=1.40.0
pydantic>=2.0.0
fastapi>=0.110.0
requests>=2.31.0
